+++ docker/scripts/R/bioconductor.R
@@ -7,8 +7,8 @@ url = "https://packagemanager.posit.co/bioconductor/__linux__/jammy/2024-10-01"
 options(BioC_mirror = "https://packagemanager.posit.co/bioconductor/2024-10-01")
 options(BIOCONDUCTOR_CONFIG_FILE = "https://packagemanager.posit.co/bioconductor/2024-10-01/config.yaml")
 
-# Configure a CRAN snapshot compatible with Bioconductor 3.20:
-options(repos = c(CRAN = url))
+# Configure a CRAN snapshot compatible with Bioconductor 3.19:
+options(repos = c(CRAN = "https://packagemanager.posit.co/cran/__linux__/jammy/2024-10-01"))
 
 # Bioconductor packages
 packages <- c(
@@ -32,8 +32,8 @@ library(devtools)
 for (package in packages) {
     tryCatch(
     {
-        devtools::install_bioc(package, mirror=url, dependencies=dependencies, upgrade=upgrade)
-        # BiocManager::install(package, version='3.12')
+        # devtools::install_bioc(package, mirror=url, dependencies=dependencies, upgrade=upgrade)
+        BiocManager::install(package, version='3.19')
     },
     error = function(e) {
         cat("Error occurred during package installation:\n")